Refactor api.py: extract Research and Presentations routes (#376)
Continues the api.py split started in refactor/api-routes-split.

- Research pipeline routes (13 endpoints) → src/pinky_daemon/routes/research.py
- Presentations + templates + share viewer + /render/pdf (16 endpoints)
  → src/pinky_daemon/routes/presentations.py

Both modules use the same APIRouter + set_dependencies(...) pattern as
the migration / kb / apps modules. The presentations module owns its
inline HTML/cookie helpers (_build_public_viewer, _build_password_gate,
_make_pres_cookie_token, _verify_pres_cookie_token, _build_generate_prompt)
and a local _session_secret() reader, which is the same shape as
routes/apps.py — there are now two copies of _session_secret(); a
follow-up can hoist it into a shared routes/_helpers.py if we don't
fold it into ctx instead.

Note: /presentations/generate calls agents.get_agent(...) which is not
a method on AgentRegistry (only .get(name) exists). That call would
raise AttributeError and 500 the request. Preserved verbatim as a
latent bug — out of scope for a behavior-preserving refactor; flagging
here so it can be fixed in a separate PR.

api.py: 9,949 → 9,420 lines (-529 lines, -5.3%)
Cumulative since main: 11,138 → 9,420 (-1,718 lines, -15.4%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
